make simulator an interface - no binary target
