--- src/simulator/simulator.cpp
@@ -1,17 +1,0 @@
-#include "simulator.hpp"
-
-
-namespace PHARE
-{
-std::unique_ptr<PHARE::ISimulator> getSimulator(std::shared_ptr<PHARE::amr::Hierarchy>& hierarchy)
-{
-    PHARE::initializer::PHAREDict const& theDict
-        = PHARE::initializer::PHAREDictHandler::INSTANCE().dict();
-    auto dim           = theDict["simulation"]["dimension"].template to<int>();
-    auto interpOrder   = theDict["simulation"]["interp_order"].template to<int>();
-    auto nbRefinedPart = theDict["simulation"]["refined_particle_nbr"].template to<int>();
-
-    return core::makeAtRuntime<SimulatorMaker>(dim, interpOrder, nbRefinedPart,
-                                               SimulatorMaker{hierarchy});
-}
-} /* namespace PHARE */
